+++ ldap-account-management/src/test/java/org/georchestra/ds/users/AccountDaoImplIT.java
@@ -1,9 +1,15 @@
 package org.georchestra.ds.users;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
 import java.util.Arrays;
 import java.util.List;
+
+import javax.naming.NamingException;
+import javax.naming.directory.Attributes;
+import javax.naming.directory.BasicAttribute;
+import javax.naming.ldap.LdapName;
 
 import org.georchestra.testcontainers.ldap.GeorchestraLdapContainer;
 import org.junit.After;
@@ -12,6 +18,7 @@ import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.ldap.core.ContextSource;
 import org.springframework.ldap.core.DirContextOperations;
 import org.springframework.ldap.core.LdapTemplate;
 import org.springframework.test.context.ContextConfiguration;
@@ -55,5 +62,77 @@ public class AccountDaoImplIT {
         account.setDescription("This is a new desc");
         accountDao.update(account, "test");
         assertTrue(Arrays.asList(dco.getStringAttributes("objectClass")).contains("dcObject"));
+    }
+
+    @Test
+    public void testBlankFields_issues_1086_1096() throws Exception {
+        Account testadminAc = accountDao.findByUID("testadmin");
+        accountDao.update(testadminAc, "admin");
+
+        ContextSource contextSource = ldapTemplate.getContextSource();
+        Attributes attrs = contextSource.getReadWriteContext().getAttributes(new LdapName("uid=testadmin,ou=users"));
+
+        boolean hasStillUserPassword = attrs.get("userPassword") != null;
+
+        accountDao.update(testadminAc, "testadmin");
+
+        assertTrue("No userPassword found for testadmin, expected one", hasStillUserPassword);
+    }
+
+    @Test
+    public void testUpdateAcountAccount() throws Exception {
+        Account testadminAc = accountDao.findByUID("testadmin");
+
+        Account newTestAdminAc = AccountFactory.create(testadminAc);
+        assertEquals(newTestAdminAc.getUid(), testadminAc.getUid());
+
+        newTestAdminAc.setUid("testadminblah");
+
+        accountDao.update(testadminAc, newTestAdminAc, "testadmin");
+
+        ContextSource contextSource = ldapTemplate.getContextSource();
+        Attributes attrs = contextSource.getReadWriteContext()
+                .getAttributes(new LdapName("uid=testadminblah,ou=users"));
+        Object o = attrs.get("uid");
+        boolean correctlyrenamed = ((BasicAttribute) o).get(0).toString().equals("testadminblah");
+        boolean encounteredNamingEx = false;
+        try {
+            Attributes oldAttrs = contextSource.getReadWriteContext()
+                    .getAttributes(new LdapName("uid=testadmin,ou=users"));
+        } catch (NamingException e) {
+            encounteredNamingEx = true;
+        }
+
+        // restoring testadmin in its initial state
+        accountDao.update(newTestAdminAc, testadminAc, "testadmin");
+
+        assertTrue("Was able to find testadmin back (found some attributes), none expected", encounteredNamingEx);
+        assertTrue("Wrong uid encountered (found " + o.toString() + " instead of testadminblah", correctlyrenamed);
+
+    }
+
+    @Test
+    public void pendingUserExists() throws Exception {
+        assertTrue(accountDao.exist("testpendinguser"));
+    }
+
+    @Test
+    public void findPendingUser() throws Exception {
+        Account testpending = accountDao.findByUID("testpendinguser");
+
+        assertEquals("testpendinguser", testpending.getUid());
+        assertTrue(testpending.isPending());
+    }
+
+    @Test
+    public void getPasswordType() throws Exception {
+        Account testpending = accountDao.findByUID("testadmin");
+        assertEquals(PasswordType.SHA, testpending.getPasswordType());
+    }
+
+    @Test
+    public void getPasswordTypePendingUser() throws Exception {
+        Account testpending = accountDao.findByUID("testpendinguser");
+        assertEquals(PasswordType.SHA, testpending.getPasswordType());
     }
 }
